chore: refactor fga retriever location and minor cleanup
