+++ packages/ai-langchain/src/FGA/FGAAuthorizer.ts
@@ -7,7 +7,7 @@ import { ToolRunnableConfig } from "@langchain/core/tools";
 export type ZodObjectAny = z.ZodObject<any, any, any, any>;
 
 /**
- * The FGAAuthorizer class implements the FGA authorization control for a Vercel AI tool.
+ * The FGAAuthorizer class implements the FGA authorization control for a LangChain AI tool.
  *
  * This class extends the FGAAuthorizerBase and provides a method to build a tool authorizer
  * that protects the tool execution using FGA.
