+++ packages/ai-langchain/test/index.test.ts
@@ -9,7 +9,7 @@ import {
   OpenFgaClient,
 } from "@openfga/sdk";
 
-import { FGARetriever } from "../src/retrievers/fga-retriever";
+import { FGARetriever } from "../src/FGA/fga-retriever";
 
 describe("FGARetriever", () => {
   process.env.FGA_CLIENT_ID = "client-id";
