+++ packages/ai-llamaindex/README.md
@@ -4,12 +4,23 @@ This package integrates [LlamaIndex](https://ts.llamaindex.ai/) with [Auth0 AI](
 
 `@auth0/ai-llamaindex` is an SDK for building secure AI-powered applications using [Auth0](https://www.auth0.ai/), [Okta FGA](https://docs.fga.dev/) and [LlamaIndex](https://ts.llamaindex.ai/).
 
-This SDK provides Okta FGA as a [retriever](https://docs.llamaindex.ai/en/stable/module_guides/querying/retriever/) for RAG applications. The retriever allows filtering documents based on access control checks defined in Okta FGA. This retriever performs batch checks on retrieved documents, returning only the ones that pass the specified access criteria.
+## Features
+
+### FGA Retriever
+
+Provides Okta FGA as a [retriever](https://docs.llamaindex.ai/en/stable/module_guides/querying/retriever/) for RAG applications. The retriever allows filtering documents based on access control checks defined in Okta FGA. This retriever performs batch checks on retrieved documents, returning only the ones that pass the specified access criteria.
+
+### FGA Authorizer
+
+Provides Okta FGA as a tool authorizer that protects the tool execution with FGA.
+
+### Async Authorizer
+
+Provides Async User Authorizer using [Client Initiated Backchannel Authentication (CIBA)](https://openid.net/specs/openid-client-initiated-backchannel-authentication-core-1_0.html).
 
 ## Install
 
-> [!WARNING] 
-> `@auth0/ai-llamaindex` is currently under development and it is not intended to be used in production, and therefore has no official support.
+> [!WARNING] > `@auth0/ai-llamaindex` is currently under development and it is not intended to be used in production, and therefore has no official support.
 
 ```
 $ npm install @auth0/ai-llamaindex
@@ -17,7 +28,7 @@ $ npm install @auth0/ai-llamaindex
 
 ## Usage
 
-Example [RAG Application](../../examples/llamaindex/retrievers-with-fga).
+Example [RAG Application](../../examples/authorization-for-rag/llamaindex).
 
 Create a Retriever instance using the `FGAReranker.create` method.
