+++ packages/ai-llamaindex/src/FGA/FGAAuthorizer.ts
@@ -5,7 +5,7 @@ import { FGAAuthorizerBase } from "@auth0/ai/FGA";
 export type LlamaToolHandler<T> = (input: T) => JSONValue | Promise<JSONValue>;
 
 /**
- * The FGAAuthorizer class implements the FGA authorization control for a Vercel AI tool.
+ * The FGAAuthorizer class implements the FGA authorization control for a LlamaIndex AI tool.
  *
  * This class extends the FGAAuthorizerBase and provides a method to build a tool authorizer
  * that protects the tool execution using FGA.
